src/requirements.txt
schedule==0.6.0
todoist-python==8.1.2
PyGithub==1.53
sqlalchemy==1.3.19